Fix Interact/BlockPlace/Break Events not being called, fix all errors!

+++ src/main/java/cn/nukkit/network/protocol/BookEditPacket.java
@@ -61,7 +61,7 @@ public class BookEditPacket extends DataPacket {
 
     @Override
     public void encode() {
-        this.putByte(this.type);
+        this.putByte((byte) this.type);
         this.putByte((byte) this.inventorySlot);
 
         switch (this.type) {
